requirements.txt
flask
python-dotenv
pylint
flask_jwt_extended
flask_jwt